track rho trading open-interest
